@@ -16,8 +16,8 @@
     "classmap": ["./src"]
   },
   "require": {
-    "php": ">=7.3",
-    "mbeym/field-mapping-input": "7.*"
+    "php": "^7.4 || ^8.0",
+    "mbeym/field-mapping-input": "8.*"
   },
   "require-dev": {
     "roave/security-advisories": "dev-latest",
